Extract common code out of clj/cljs base plugins
Use a shared common plugin to provide the basic necesseties. This should
allow someone to use the CLJS plugin without CLJ. Not a huge need, but
makes it a little cleaner internally.
